09-04-2023

+++ EMS-main-master/employee/src/main/java/com/ems/emploee/Repository/registeredemployeerepository.java
@@ -1,13 +1,15 @@
 package com.ems.emploee.Repository;
 
 import com.ems.emploee.entity.Employee;
+import com.ems.emploee.model.EmployeeModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
+import java.util.ArrayList;
 import java.util.List;
 
-public interface registeredemployeerepository extends JpaRepository<Employee,Integer> {
+public interface registeredemployeerepository extends JpaRepository<Employee,Long> {
 
 
-    }
+}
